# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Odds Dropper API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Sport {
  slug: string;
  name: string;
  icon: string;
  eventCount: number;
}

export interface League {
  slug: string;
  name: string;
  sport: string;
  country: string;
  eventCount: number;
}

export type OddsLineDirection =
  (typeof OddsLineDirection)[keyof typeof OddsLineDirection];

export const OddsLineDirection = {
  drop: "drop",
  rise: "rise",
  stable: "stable",
} as const;

export interface OddsLine {
  selection: string;
  openingOdds: number;
  currentOdds: number;
  changePercent: number;
  changeAbsolute: number;
  direction: OddsLineDirection;
}

export type OddsEventMarketType =
  (typeof OddsEventMarketType)[keyof typeof OddsEventMarketType];

export const OddsEventMarketType = {
  moneyline: "moneyline",
  spread: "spread",
  total: "total",
  asian_handicap: "asian_handicap",
} as const;

export interface OddsEvent {
  id: string;
  homeTeam: string;
  awayTeam: string;
  sport: string;
  league: string;
  leagueName: string;
  commenceTime: string;
  marketType: OddsEventMarketType;
  lines: OddsLine[];
  biggestDrop: number;
  biggestRise: number;
  newDropAt?: string | null;
  lastUpdated: string;
}

export interface OddsMovement {
  timestamp: string;
  odds: number;
  selection: string;
}

export type OddsEventDetail = OddsEvent & {
  movements: OddsMovement[];
};

export interface OddsSummary {
  totalEvents: number;
  dropsCount: number;
  risesCount: number;
  avgDropPercent: number;
  biggestDrop: number;
  biggestRise: number;
  lastUpdated: string;
}

export type GetOddsDropsParams = {
  /**
   * Filter by sport slug
   */
  sport?: string;
  /**
   * Filter by league slug
   */
  league?: string;
  /**
   * Minimum percentage drop to include
   */
  minDrop?: number;
  /**
   * Filter by movement direction
   */
  direction?: GetOddsDropsDirection;
};

export type GetOddsDropsDirection =
  (typeof GetOddsDropsDirection)[keyof typeof GetOddsDropsDirection];

export const GetOddsDropsDirection = {
  drop: "drop",
  rise: "rise",
  all: "all",
} as const;