# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Odds Dropper API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /odds/drops:
    get:
      operationId: getOddsDrops
      tags: [odds]
      summary: Get odds drops
      description: Returns events with significant odds movements
      parameters:
        - name: sport
          in: query
          required: false
          schema:
            type: string
          description: Filter by sport slug
        - name: league
          in: query
          required: false
          schema:
            type: string
          description: Filter by league slug
        - name: minDrop
          in: query
          required: false
          schema:
            type: number
          description: Minimum percentage drop to include
        - name: direction
          in: query
          required: false
          schema:
            type: string
            enum: [drop, rise, all]
          description: Filter by movement direction
      responses:
        "200":
          description: List of odds drop events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OddsEvent"
  /odds/drops/{id}:
    get:
      operationId: getOddsDropById
      tags: [odds]
      summary: Get single odds drop event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Odds event detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OddsEventDetail"
        "404":
          description: Not found
...
components:
  schemas:
...
    OddsLine:
      type: object
      properties:
        selection:
          type: string
        openingOdds:
          type: number
        currentOdds:
          type: number
        changePercent:
          type: number
        changeAbsolute:
          type: number
        direction:
          type: string
          enum: [drop, rise, stable]
      required:
        - selection
        - openingOdds
        - currentOdds
        - changePercent
        - changeAbsolute
        - direction
    OddsEvent:
      type: object
      properties:
        id:
          type: string
        homeTeam:
          type: string
        awayTeam:
          type: string
        sport:
          type: string
        league:
          type: string
        leagueName:
          type: string
        commenceTime:
          type: string
          format: date-time
        marketType:
          type: string
          enum: [moneyline, spread, total, asian_handicap]
        lines:
          type: array
          items:
            $ref: "#/components/schemas/OddsLine"
        biggestDrop:
          type: number
        biggestRise:
          type: number
        newDropAt:
          type: string
          format: date-time
          nullable: true
        lastUpdated:
          type: string
          format: date-time
      required:
        - id
        - homeTeam
        - awayTeam
        - sport
        - league
        - leagueName
        - commenceTime
        - marketType
        - lines
        - biggestDrop
        - biggestRise
        - lastUpdated
    OddsMovement:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        odds:
          type: number
        selection:
          type: string
      required:
        - timestamp
        - odds
        - selection
    OddsEventDetail:
      type: object
      allOf:
        - $ref: "#/components/schemas/OddsEvent"
        - type: object
          properties:
            movements:
              type: array
              items:
                $ref: "#/components/schemas/OddsMovement"
          required:
            - movements
...