# lib/api-zod/src/generated/types/oddsEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Odds Dropper API
 * OpenAPI spec version: 0.1.0
 */
import type { OddsEventMarketType } from "./oddsEventMarketType";
import type { OddsLine } from "./oddsLine";

export interface OddsEvent {
  id: string;
  homeTeam: string;
  awayTeam: string;
  sport: string;
  league: string;
  leagueName: string;
  commenceTime: Date;
  marketType: OddsEventMarketType;
  lines: OddsLine[];
  biggestDrop: number;
  biggestRise: number;
  newDropAt?: Date | null;
  lastUpdated: Date;
}