# lib/api-zod/src/generated/types/oddsMovement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Odds Dropper API
 * OpenAPI spec version: 0.1.0
 */

export interface OddsMovement {
  timestamp: Date;
  odds: number;
  selection: string;
  limit?: number | null;
}